+++ sdk/resourcemanager/securityinsights/armsecurityinsights/CHANGELOG.md
@@ -1,5 +1,89 @@
 # Release History
 
+## 0.4.0 (2022-05-15)
+### Breaking Changes
+
+- Type of `GetInsightsResultsMetadata.Errors` has been changed from `[]*GetInsightsError` to `[]*GetInsightsErrorKind`
+- Const `SKUKindPerGB` has been removed
+- Const `SKUKindCapacityReservation` has been removed
+- Const `GetInsightsErrorKindInsight` has been removed
+- Function `PossibleSKUKindValues` has been removed
+- Function `PossibleGetInsightsErrorKindValues` has been removed
+- Struct `GetInsightsError` has been removed
+- Struct `SKU` has been removed
+
+### Features Added
+
+- New const `AutomationRulePropertyChangedConditionSupportedPropertyTypeIncidentSeverity`
+- New const `AutomationRulePropertyChangedConditionSupportedPropertyTypeIncidentOwner`
+- New const `AutomationRulePropertyChangedConditionSupportedPropertyTypeIncidentStatus`
+- New const `AutomationRulePropertyArrayChangedConditionSupportedArrayTypeTactics`
+- New const `AutomationRulePropertyArrayChangedConditionSupportedArrayTypeLabels`
+- New const `EntityTimelineKindAnomaly`
+- New const `ConditionTypePropertyChanged`
+- New const `SettingsStatusProduction`
+- New const `SettingsStatusFlighting`
+- New const `GetInsightsErrorInsight`
+- New const `ConditionTypePropertyArrayChanged`
+- New const `AutomationRulePropertyChangedConditionSupportedChangedTypeChangedFrom`
+- New const `AutomationRulePropertyArrayChangedConditionSupportedChangeTypeAdded`
+- New const `AutomationRulePropertyArrayChangedConditionSupportedArrayTypeComments`
+- New const `AutomationRulePropertyArrayChangedConditionSupportedArrayTypeAlerts`
+- New const `AutomationRulePropertyChangedConditionSupportedChangedTypeChangedTo`
+- New const `SecurityMLAnalyticsSettingsKindAnomaly`
+- New const `TriggersWhenUpdated`
+- New function `PossibleAutomationRulePropertyChangedConditionSupportedPropertyTypeValues() []AutomationRulePropertyChangedConditionSupportedPropertyType`
+- New function `*PropertyChangedConditionProperties.GetAutomationRuleCondition() *AutomationRuleCondition`
+- New function `PossibleSettingsStatusValues() []SettingsStatus`
+- New function `AnomalySecurityMLAnalyticsSettingsProperties.MarshalJSON() ([]byte, error)`
+- New function `AnomalySecurityMLAnalyticsSettings.MarshalJSON() ([]byte, error)`
+- New function `PossibleAutomationRulePropertyChangedConditionSupportedChangedTypeValues() []AutomationRulePropertyChangedConditionSupportedChangedType`
+- New function `PossibleSecurityMLAnalyticsSettingsKindValues() []SecurityMLAnalyticsSettingsKind`
+- New function `*AnomalyTimelineItem.GetEntityTimelineItem() *EntityTimelineItem`
+- New function `PropertyArrayChangedConditionProperties.MarshalJSON() ([]byte, error)`
+- New function `*AnomalyTimelineItem.UnmarshalJSON([]byte) error`
+- New function `*PropertyChangedConditionProperties.UnmarshalJSON([]byte) error`
+- New function `*PropertyArrayChangedConditionProperties.UnmarshalJSON([]byte) error`
+- New function `AnomalyTimelineItem.MarshalJSON() ([]byte, error)`
+- New function `SecurityMLAnalyticsSettingsDataSource.MarshalJSON() ([]byte, error)`
+- New function `*AnomalySecurityMLAnalyticsSettings.UnmarshalJSON([]byte) error`
+- New function `PossibleGetInsightsErrorValues() []GetInsightsError`
+- New function `*PropertyArrayChangedConditionProperties.GetAutomationRuleCondition() *AutomationRuleCondition`
+- New function `SecurityMLAnalyticsSettingsList.MarshalJSON() ([]byte, error)`
+- New function `PossibleAutomationRulePropertyArrayChangedConditionSupportedArrayTypeValues() []AutomationRulePropertyArrayChangedConditionSupportedArrayType`
+- New function `*AnomalySecurityMLAnalyticsSettings.GetSecurityMLAnalyticsSetting() *SecurityMLAnalyticsSetting`
+- New function `*SecurityMLAnalyticsSettingsClientGetResponse.UnmarshalJSON([]byte) error`
+- New function `*SecurityMLAnalyticsSettingsList.UnmarshalJSON([]byte) error`
+- New function `PossibleAutomationRulePropertyArrayChangedConditionSupportedChangeTypeValues() []AutomationRulePropertyArrayChangedConditionSupportedChangeType`
+- New function `*SecurityMLAnalyticsSettingsClientCreateOrUpdateResponse.UnmarshalJSON([]byte) error`
+- New function `AutomationRulePropertyValuesChangedCondition.MarshalJSON() ([]byte, error)`
+- New function `*AnomalySecurityMLAnalyticsSettingsProperties.UnmarshalJSON([]byte) error`
+- New function `*SecurityMLAnalyticsSetting.GetSecurityMLAnalyticsSetting() *SecurityMLAnalyticsSetting`
+- New function `PropertyChangedConditionProperties.MarshalJSON() ([]byte, error)`
+- New struct `AnomalySecurityMLAnalyticsSettings`
+- New struct `AnomalySecurityMLAnalyticsSettingsProperties`
+- New struct `AnomalyTimelineItem`
+- New struct `AutomationRulePropertyArrayChangedValuesCondition`
+- New struct `AutomationRulePropertyValuesChangedCondition`
+- New struct `GetInsightsErrorKind`
+- New struct `PropertyArrayChangedConditionProperties`
+- New struct `PropertyChangedConditionProperties`
+- New struct `SecurityMLAnalyticsSetting`
+- New struct `SecurityMLAnalyticsSettingsClientCreateOrUpdateOptions`
+- New struct `SecurityMLAnalyticsSettingsClientCreateOrUpdateResponse`
+- New struct `SecurityMLAnalyticsSettingsClientDeleteOptions`
+- New struct `SecurityMLAnalyticsSettingsClientDeleteResponse`
+- New struct `SecurityMLAnalyticsSettingsClientGetOptions`
+- New struct `SecurityMLAnalyticsSettingsClientGetResponse`
+- New struct `SecurityMLAnalyticsSettingsClientListOptions`
+- New struct `SecurityMLAnalyticsSettingsClientListResponse`
+- New struct `SecurityMLAnalyticsSettingsDataSource`
+- New struct `SecurityMLAnalyticsSettingsList`
+- New field `DataCollectionRuleImmutableID` in struct `DataConnectorConnectBody`
+- New field `DataCollectionEndpoint` in struct `DataConnectorConnectBody`
+- New field `OutputStream` in struct `DataConnectorConnectBody`
+
+
 ## 0.3.0 (2022-04-18)
 ### Breaking Changes